// backend/src/mysql/models/Institution.ts
export class Institution{
    idInstitucije: number = 0;
    nazivInstSrp: string = "";
    nazivInstEng: string = "";
    nazivSkrInst: string = "";
    adresaInst: string = "";
    gradInst: string = "";
    ovlascenoLiceInst: string = "";
    telefonInst: string = "";
    sifraInstitucije: string = "";
}